- Refactor to build in the various configurations.   - The whole QDQ setup needs a rethink at some point as it's currently spread across too many places (framework, optimizer, base providers lib, EP specific providers lib) - move NodeGroup to framework/node_unit.h and ValidateNodeGroupQDQNodes to NodeGroup::CanCreateNodeGroup so it's in the framework lib as it's used by NodeUnit - move GetAllNodeUnits to optimizer   - doesn't quite belong there but this works will all the current EPs that use it.
