+++ onnxruntime/core/optimizer/qdq_transformer/selectors_actions/shared/utils.cc
@@ -13,6 +13,7 @@
 #include <core/providers/common.h>
 
 #include "core/optimizer/qdq_transformer/selectors_actions/qdq_selectors.h"
+#include "core/optimizer/qdq_transformer/selectors_actions/shared/utils.h"
 
 namespace onnxruntime {
 namespace QDQ {
@@ -324,64 +325,48 @@ std::vector<NodeGroup> SelectorManager::GetQDQSelections(const GraphViewer& grap
   return qdq_selections;
 }
 
-Status ValidateNodeGroupQDQNodes(const GraphViewer& graph_viewer,
-                                 const Node& target_node,
-                                 gsl::span<const Node* const> dq_nodes,
-                                 gsl::span<const Node* const> q_nodes) {
-  // Within a QDQ node group, a target node input is the only consumer of each DQ.
-  // This should have been ensured by the EnsureUniqueDQForNodeUnit graph transformer, but other graph modifications
-  // may have happened since. Verify that this is still true.
-  for (const auto* dq_node : dq_nodes) {
-    const bool dq_produces_graph_output = graph_viewer.NodeProducesGraphOutput(*dq_node);
-    ORT_RETURN_IF(dq_produces_graph_output,
-                  "QDQ node group cannot have DQ node that produces a graph output. DQ node: ", dq_node->Name(),
-                  ", target node: ", target_node.Name());
+std::pair<std::vector<std::unique_ptr<NodeUnit>>, std::unordered_map<const Node*, const NodeUnit*>>
+GetAllNodeUnits(const GraphViewer& graph_viewer) {
+  std::vector<std::unique_ptr<NodeUnit>> node_unit_holder;
+  std::unordered_map<const Node*, const NodeUnit*> node_unit_map;
 
-    const bool dq_has_single_output_edge_to_target =
-        dq_node->GetOutputEdgesCount() == 1 &&
-        dq_node->OutputEdgesBegin()->GetNode().Index() == target_node.Index();
-    ORT_RETURN_IF_NOT(dq_has_single_output_edge_to_target,
-                      "QDQ node group cannot have DQ that doesn't have a single output edge to the target node. "
-                      "DQ node: ",
-                      dq_node->Name(), ", target node: ", target_node.Name());
-  }
-
-  // an output from the target node can have either Q consumers or direct consumers. it cannot have both.
-  // this must be checked on a per output basis.
-  // NOTE: rules about the target node not producing a graph output must be checked by the selector as it's operator
-  // dependent.
-  // e.g. TopK produces values and indices. The indices output won't be quantized, so even if we replace the TopK QDQ
-  // node group with a quantized TopK, an int64_t indices value will be produced and can provide a graph output.
-  if (!q_nodes.empty()) {
-    auto cur_edge = target_node.OutputEdgesBegin();
-    auto end_edge = target_node.OutputEdgesEnd();
-    std::vector<const Node*> output_consumers(target_node.OutputDefs().size(), nullptr);
-
-    for (; cur_edge != end_edge; ++cur_edge) {
-      auto output_idx = cur_edge->GetSrcArgIndex();
-      const Node& this_consumer = cur_edge->GetNode();
-      const Node* existing_consumer = output_consumers[output_idx];
-
-      if (existing_consumer != nullptr) {
-        // another edge for this output. either both are Q or both are not.
-        bool valid = true;
-        if (existing_consumer->OpType() == "QuantizeLinear") {
-          valid = this_consumer.OpType() == "QuantizeLinear";
-        } else {
-          valid = this_consumer.OpType() != "QuantizeLinear";
-        }
-
-        ORT_RETURN_IF_NOT(valid,
-                          "QDQ node group cannot have an output from the target node being consumed by a Q node and "
-                          "a non-Q node. target node: ",
-                          target_node.Name());
-      } else {
-        output_consumers[output_idx] = &this_consumer;
-      }
+  const auto add_node_unit_to_map = [&](const std::vector<NodeIndex>& node_indices, const NodeUnit* node_unit) {
+    for (const auto& node_idx : node_indices) {
+      const auto* node = graph_viewer.GetNode(node_idx);
+      node_unit_map.insert({node, node_unit});
     }
+  };
+
+  // Get QDQ NodeUnits first
+  QDQ::SelectorManager selector_mgr;
+  const auto qdq_selections = selector_mgr.GetQDQSelections(graph_viewer);
+
+  for (const auto& qdq_selection : qdq_selections) {
+    auto qdq_unit = std::make_unique<NodeUnit>(graph_viewer, qdq_selection);
+
+    // Fill the node to node_unit map for all nodes in the QDQ Group
+    add_node_unit_to_map(qdq_selection.dq_nodes, qdq_unit.get());
+    add_node_unit_to_map(qdq_selection.q_nodes, qdq_unit.get());
+    add_node_unit_to_map({qdq_selection.target_node}, qdq_unit.get());
+
+    node_unit_holder.push_back(std::move(qdq_unit));
   }
 
-  return Status::OK();
+  // Get the left over SingleNode NodeUnits
+  const auto& node_indices = graph_viewer.GetNodesInTopologicalOrder();
+  for (const auto node_idx : node_indices) {
+    const auto* node(graph_viewer.GetNode(node_idx));
+
+    // This is already part of a QDQ NodeUnit
+    if (node_unit_map.find(node) != node_unit_map.cend())
+      continue;
+
+    auto node_unit = std::make_unique<NodeUnit>(*node);
+    node_unit_map[node] = node_unit.get();
+    node_unit_holder.push_back(std::move(node_unit));
+  }
+
+  return std::make_pair(std::move(node_unit_holder), std::move(node_unit_map));
 }
 
 }  // namespace QDQ
